Refactor ApplicationService moving it to own package and creating two interfaces.
